requirements.txt
asgiref==3.6.0
Django==4.1.7
django-filter==22.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
gunicorn==20.1.0
phonenumberslite==8.13.6
python-decouple==3.7
pytz==2022.7.1
sqlparse==0.4.3
